requirements.txt
Flask~=3.0.0
requests==2.31.0
youtube-transcript-api==0.6.2
pytube==15.0.0
python-dotenv==1.0.0